pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Leandro Franco <leandro00fr@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = true

[tool.poetry.dependencies]
python = "^3.12"
ffmpeg-python = "^0.2.0"
python-dotenv = "^1.0.1"
colorama = "^0.4.6"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = 'app.main:main'

[tool.ruff]
line-length = 100

select = [
  "E",
  "F",
  "UP",
  "I",
]
